Configue correct extended_key_usage on backup certificates
During the previous commit, the add-backup.yml ops-file was incorrectly
updated to set extended_key_usage to client_auth on the _server_
certificate. This caused the tls handshake to fail between the backup
client and backup server with the error:

 x509: certificate specifies an incompatible key usage

Now the backup server certificate is set for server_auth and the backup
client certificate is set for client_auth.

[#170295306]