streaming-mysql-backup-tool: remove test pollution in config_test
The enableMutualTLS flag was not initialized so test order could leave
this flag enabled, but run tests that assumed the default, false, value
was configured leading to unexpected failures.

Now the enableMutualTLS flag is always initialized to false in test
setup.

Additionally removed some unnecessary "global" test state as part of
cleanup while in the area.