packaging: fix xbcloud dependency in xtrabackup distributions
xbcloud previously was broken because it linked to libev which was not
installed in the correct location.

The xbcloud tool is not directly used by mysql-backup-release but is
sometimes useful as a troubleshooting tool and is not trivially
disabled. This change fixes the install so the binary functions as
intended.

[#184634391](https://www.pivotaltracker.com/story/show/184634391)